Remove duplicate cache ID check
This system was built with the intention that a
single component will hold reference to a cached image,
and will unmount before another component can take it.
In our case, we want multiple components to hold
a cached reference to an image.
Example: There could be user A's avatar in the contacts
screen, and also in the pub screen (tabs don't get unloaded).